Clean up SignatureHelpProvider and other code clean up
