+++ src/Features/VisualBasic/Portable/SignatureHelp/Providers/RaiseEventStatementSignatureHelpProvider.vb
@@ -73,10 +73,11 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.SignatureHelp.Providers
                 semanticModel.LookupSymbols(raiseEventStatement.SpanStart, containingType, raiseEventStatement.Name.Identifier.ValueText))
 
             Dim symbolDisplayService = document.Project.LanguageServices.GetService(Of ISymbolDisplayService)()
-            Dim allowedEvents = events.Where(Function(s) s.Kind = SymbolKind.Event AndAlso s.ContainingType Is containingType).
-                                       Cast(Of IEventSymbol)().
-                                       FilterToVisibleAndBrowsableSymbolsAndNotUnsafeSymbols(document.ShouldHideAdvancedMembers(), semanticModel.Compilation).
-                                       Sort(symbolDisplayService, semanticModel, raiseEventStatement.SpanStart)
+            Dim allowedEvents = events _
+                .Where(Function(s) s.Kind = SymbolKind.Event AndAlso s.ContainingType Is containingType) _
+                .Cast(Of IEventSymbol)() _
+                .FilterToVisibleAndBrowsableSymbolsAndNotUnsafeSymbols(document.ShouldHideAdvancedMembers(), semanticModel.Compilation) _
+                .Sort(symbolDisplayService, semanticModel, raiseEventStatement.SpanStart)
 
             Dim anonymousTypeDisplayService = document.Project.LanguageServices.GetService(Of IAnonymousTypeDisplayService)()
             Dim documentationCommentFormattingService = document.Project.LanguageServices.GetService(Of IDocumentationCommentFormattingService)()
